backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
python-socketio
pydantic-settings
pydantic[email]
email-validator
